Refactors BosStateManager into User and Conversation components
